apps/classifier/model-service/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
transformers==4.38.1
torch==2.2.0
torchvision==0.17.0
Pillow==10.2.0
httpx==0.26.0
python-multipart==0.0.9
pydantic==2.6.1
numpy==1.26.4
